Herdimmunity

+++ herdimmunity/src/main/scala/com/revature/scalawags/project3/herdimmunity/AnalysisData.scala
@@ -4,8 +4,8 @@ import com.github.nscala_time.time.Imports._
 
 case class AnalysisData(
   date: DateTime, 
-  peopleVaccinated: Double, 
-  peopleFullyVaccinated: Double, 
-  newVaccinationsSmoothed: Double, 
-  population: Double
+  peopleVaccinated: Int, 
+  peopleFullyVaccinated: Int, 
+  newVaccinationsSmoothed: Int, 
+  population: Int
 )
